pyproject.toml
[project]
name = "jurisiq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.8.3",
    "duckdb>=1.5.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "uvicorn[standard]>=0.42.0",
]
